ExerciseModule29/Module-29/src/marcos/ferrel/domain/Product.java
package marcos.ferrel.domain;

/**
 * @author marcos.ferrel
 */
public class Product {

    private Long id;

    private String pname;

    private String pcode;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getPname() {
        return pname;
    }

    public void setPname(String pname) {
        this.pname = pname;
    }

    public String getPcode() {
        return pcode;
    }

    public void setPcode(String pcode) {
        this.pcode = pcode;
    }
}
